+++ device-management-application/conanfile.py
@@ -8,7 +8,8 @@ required_conan_version = ">=1.52.0"
 
 class DeviceManagementApplicationConan(ConanFile):
     name = "deviceManagementApplication"
-    url = "https://gitlab.esperanto.ai/software/device-management-application"
+    url = "git@gitlab.com:esperantotech/software/device-management-application.git"
+    homepage = "https://gitlab.com/esperantotech/software/device-management-application"
     description = ""
     license = "Esperanto Technologies"
 
@@ -20,12 +21,6 @@ class DeviceManagementApplicationConan(ConanFile):
         "fPIC": True,
     }
 
-    scm = {
-        "type": "git",
-        "url": "git@gitlab.esperanto.ai:software/device-management-application.git",
-        "revision": "auto",
-    }
-
     python_requires = "conan-common/[>=1.1.0 <2.0.0]"
 
     def set_version(self):
@@ -35,7 +30,11 @@ class DeviceManagementApplicationConan(ConanFile):
     def config_options(self):
         if self.settings.os == "Windows":
             del self.options.fPIC
-    
+
+    def export(self):
+        register_scm_coordinates = self.python_requires["conan-common"].module.register_scm_coordinates
+        register_scm_coordinates(self)
+
     def configure(self):
         check_req_min_cppstd = self.python_requires["conan-common"].module.check_req_min_cppstd
         check_req_min_cppstd(self, "17")
@@ -51,7 +50,15 @@ class DeviceManagementApplicationConan(ConanFile):
 
         self.requires("fmt/7.1.3")
         self.requires("glog/0.4.0")
-    
+
+    def export_sources(self):
+        copy_sources_if_scm_dirty = self.python_requires["conan-common"].module.copy_sources_if_scm_dirty
+        copy_sources_if_scm_dirty(self)
+
+    def source(self):
+        get_sources_if_scm_pristine = self.python_requires["conan-common"].module.get_sources_if_scm_pristine
+        get_sources_if_scm_pristine(self)
+
     def generate(self):
         tc = CMakeToolchain(self)
         tc.variables["CMAKE_INSTALL_LIBDIR"] = "lib"